Assignment 1
Assignment 1: You are required to develop a small system using MAVEN and Java programming language. The system should be able to:

Read data from a table of a website. Refer to the link below for the testing purpose (see Trivia table). https://ms.wikipedia.org/wiki/Malaysia Convert/write the data from the table to a new MS Excel file.